+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Added `data-modal-open` attribute to root document when modal is opened
+* Added check for `data-overlay-open` attribute on root document to prevent removal of `inert` attribute if an overlay is also open
 
 3.3.0 - (June 19, 2019)
 ------------------